Robust, well-done system for handling cartridges
This. Took. WAY too long.
-Revamp to the entire system of title handling. Cartridges are now always placed in the first position in the title IDs vector. They are always selected in the Active Titles menu and cannot be turned off. The cartridges are updated by subscribing to the cartridge insertion/ejection SRV events, which triggers a function to let the main thread know that it needs to be updated. (I tried to do this from the SRV thread itself, but several days in I still could not solve a crash and figured it wasn't worth the time).
This was originally handled by calling the FS command for checking cartridge insertion constantly, but seeing as FS has a ridiculous amount of overhead (needing to talk to arm9) it lagged the program to an unacceptable point. So I had to put in several days to revamp it using SRV taking several days, then dealing with the absurd and seemingly random crashes for more days on end... Ugh. This was extremely stupid and I'm glad it's done.