Initial setup cleanup, SS 2.x migrator
Hopefully all the changes to the initial setup menu aren't terribly broken, but they *seem* to be working.
Now with 100% less segfaults by attempting to access something in the all titles vector that hasn't been loaded yet!
Now with 10% more std functionality! std::tuple is awesome.
Now with some very cleaned up initial setup code! Much prettier and looks better on the screen.
This commit adds a lot of moving parts. If something ends up terribly broken, this is the commit to blame.