Add "dark mode"
This inverts the colors of most UI aspects to create a dark theme. Should be nice for most of the dark theme supremacists.
Also improves code sanity in title selections. TevOpAlpha should configure the ops for both sources. The default values for C3D_TexEnvOpAlpha would do this anyway, but that's a terrible thing to rely on.