New config settings
A setting for dark mode, which really should have been done in the last commit.
A setting for disabling error calls, and a setting for disabling the updater. I can't believe I'm doing this; they have their purpose, but that purpose is going to be shrouded by the idiot masses who don't know what they're doing.