New (animated!) logo!
Import the tutorial and dark mode buttons (temporary). The real meat of this commit is the fun animation of the new logo.
It functions by taking an alpha map based on the rim of the moon (where the colors go) and steadily moving a rainbow texture through it. Extremely cool, but it showcases a huge design problem with sDraw... I needed a new vertex shader to pass in the texcoords, but it's *extremely hard* to add one without any sort of shader manager! When revamp time comes around, I have a couple ideas to fix this (and all the other design issues- they'll be all taken care of in a huge ripping out of the floorboards...)