Fix the main menu highlighter breaking in an edge case
It would only trigger if a progress bar was used, so I stared at it for a while, took a look at the pixels being discarded, and laughed because the bug was silly
It's a good idea to *clear your buffers* if you rely on them being used in more than one scenario. The stencil buffer wasn't being reset, and the progress bar was writing things there. So any time the highlighter algorithm looked to see if there wasn't a pixel set, it would fail in any pixel the progress bar was drawn in.
Which ended up looking like someone had just chopped a ton of pixels off near the bottom of the screen...
Also adds a large comment to explain why I'm still doing PNG conversion in August 2018.