Moderate sDraw revamp- use Tex3DS
Tex3DS is a really neat system that stores textures in native format, and even builds spritesheets and stores texcoords internally so you don't have to generate/manage the spritesheet manually. Thanks to the awesome efforts of @mtheall, this revamp was made painless with a script (included in resources) that automatically built textures from the spritesheet and the previous sdraw_stex coordinates. Furthermore, he wrote Tex3DS in the first place. Props to him!
This should lead to a significant performance increase. Not only is loading PNG files eliminated, spritesheet calculation at every single draw call (WTF???) was removed in favor of a single calculation at sdraw_stex construction time.