WIP: Vertex shader manager
This has been in the works for a while. It's a complete vertex shader manager class. The implementation itself is completely done and will be documented shortly. However, the current implementation into sDraw is very, very incomplete- it's basically hacked in as a proof of concept. Only the basic shader is used. Lots of hacks are used to get it to that point. The engine core will be seeing some areas getting rewrites to get it to a fully functional state.