Make the behavior of drawrectangle consistent
drawrectangle() should not be switching state within the core. Instead, the user should set the fragment shader for color and then call that function for vertex adding, consistent with the rest of sDraw's paradigms. This change was missed in the earlier fragment shader overhaul, and I'm betting it caused one or two rendering bugs because I was unaware of the state switch...