Add fully featured modpack downloader
This modpack downloader downloads a ZIP file and extracts it using the ZIP code written for Lua Player Plus. It then extracts it into the first free slot of the title the modpack is for, activating the title if necessary.
A tools menu button is currently in the works. The mod migration button is still currently there and manual usage of it will be removed. It was really only a placeholder since it was done automatically at first start, the only time you would ever need to use it.