Lots, lots, LOTS of changed things!
All of this work is several months old, so I don't really know the extent of what's in it. I believe SaltySD hitbox display was implemented, and furthermore the slot deleter was fixed. Also, a strange edge case with Horizon's FS module was worked around. That should be the majority of it. I still don't know if or when I'll release 3.1.