(Untested) Build System: Attempt to detect running as 3DSX or CIA at runtime
When homebrew programs are run as CIA files, argc and argv are 0. This property allows us to detect at runtime what type of update to install, instead of a cumbersome #define that needs to be *manually* added in the source. Said define could have been added to the Makefile, as I've learned, but I just didn't know how at the time. This is better anyway.